Create README.md
If you’re new to Redis, you may be wondering what exactly it is and who should use it.
##What is Redis?
Redis (which stands for REmote DIctionary Server) is an open source, in-memory, NoSQL key/value store that is primarily used as an application cache or quick-response database. It provides unrivaled speed, reliability, and performance since it saves data in memory rather than on a disk or solid-state drive (SSD).

When an app uses external data sources, the latency and throughput of such sources might be a performance barrier, especially as traffic grows and the app scales. In these instances, one method to increase performance is to store and manipulate data in memory, which is physically closer to the application.